creates abstraction for confirm variables
