+++ feature_engine/selection/drop_psi_features.py
@@ -268,6 +268,8 @@ class DropHighPSIFeatures(BaseSelector):
                 f"confirm_variables must be a boolean. Got {confirm_variables} instead."
             )
 
+        super().__init__(confirm_variables)
+
         # Check the variables before assignment.
         self.variables = _check_input_parameter_variables(variables)
 
@@ -282,7 +284,6 @@ class DropHighPSIFeatures(BaseSelector):
         self.strategy = strategy
         self.min_pct_empty_bins = min_pct_empty_bins
         self.missing_values = missing_values
-        self.confirm_variables = confirm_variables
 
     def fit(self, X: pd.DataFrame, y: pd.Series = None):
         """
@@ -300,10 +301,7 @@ class DropHighPSIFeatures(BaseSelector):
         X = _is_dataframe(X)
 
         # If required exclude variables that are not in the input dataframe
-        if self.confirm_variables:
-            self.variables_ = _filter_out_variables_not_in_dataframe(X, self.variables)
-        else:
-            self.variables_ = self.variables
+        self._confirm_variables(X)
 
         # find numerical variables or check those entered are present in the dataframe
         self.variables_ = _find_or_check_numerical_variables(X, self.variables_)
